# PERF: BaseMaskedArray.__iter__