- [x] [Tests added and passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#writing-tests) if fixing a bug or adding a new feature
- [x] All [code checks passed](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#pre-commit).
- [x] Added [type annotations](https://pandas.pydata.org/pandas-docs/dev/development/contributing_codebase.html#type-hints) to new arguments/methods/functions.
- [x] Added an entry in the latest `doc/source/whatsnew/v2.0.0.rst` file if fixing a bug or adding a new feature.

Faster iteration through EA's backed by BaseMaskedArray.

ASV added for iterating through a `Series` backed by various dtypes:

```
       before           after         ratio
     [b6736b44]       [b9193932]
                      <masked-array-iter>
-      16.8±0.9ms      5.14±0.06ms     0.31  series_methods.Iter.time_iter('Float64')
-        17.2±1ms       5.15±0.1ms     0.30  series_methods.Iter.time_iter('Int64')
-        16.8±1ms      3.50±0.02ms     0.21  series_methods.Iter.time_iter('boolean')
```